+++ src/core/api.h
@@ -1,0 +1,7 @@
+
+#ifndef P2G_CORE_API_H
+#define P2G_CORE_API_H
+
+#define LIB_PREFIX "P2Garage."
+#define MAKE_LUA_LIB_NAME(n) (LIB_PREFIX n)
+#endif
