+++ src/core/buffer.c
