+++ src/include/p2g/gs.h
@@ -5,4 +5,6 @@
 int gs_init();
 int gs_flip();
 
+int gs_lua_init(lua_State *l);
+
 #endif
