+++ src/include/p2g/script.h
@@ -1,0 +1,9 @@
+
+#ifndef SRC_LUA_H
+#define SRC_LUA_H
+
+#include <lua.h>
+
+int bind_core_libs(lua_State *l);
+
+#endif
