+++ src/main.c
@@ -12,14 +12,13 @@
 #include <kernel.h>
 #include <sifrpc.h>
 
-#include "log.h"
-
-#include "bench.h"
-#include "gs.h"
-#include "pad.h"
-#include "ps2luaprog.h"
-#include "script.h"
-#include "utils.h"
+#include <p2g/log.h>
+#include <p2g/bench.h>
+#include <p2g/gs.h>
+#include <p2g/pad.h>
+#include <p2g/ps2luaprog.h>
+#include <p2g/script.h>
+#include <p2g/utils.h>
 
 static int is_running = 1;
 
@@ -50,21 +49,6 @@ char main_script[FILE_NAME_MAX_LEN];
     while (1) {                                                                \
     }                                                                          \
   } while (0)
-
-static script_binding SCRIPT_CORE_LIBS[] = {
-    {"gs", gs_lua_init},
-    {"dma", dma_lua_init},
-    {"pad", pad_lua_init},
-    {"buffer", drawlua_init},
-    {"log", loglua_init},
-    {"tga", lua_tga_init},
-    {"slotlist", slot_list_lua_init},
-    {"draw2d", draw2d_lua_init},
-    {"math_vec2", vec2lua_init},
-    {"math_vec3", vec3lua_init},
-    {"math_mat3", mat3lua_init},
-    {"math_misc", floatmath_init},
-};
 
 void core_error(const char *msg) {
   is_running = 0;
@@ -122,23 +106,6 @@ static int runfile(lua_State *l, const char *fname) {
   BENCH_INFO(runfile_time, " - total run time %f");
 
   return 0;
-}
-
-int bind_core_libs(lua_State *l) {
-  int num_libs = sizeof(SCRIPT_CORE_LIBS) / sizeof(script_binding);
-  info("initializing %d core libraries", num_libs);
-  lua_createtable(l, 0, num_libs);
-  for (int i = 0; i < num_libs; i++) {
-    script_binding *b = &SCRIPT_CORE_LIBS[i];
-    trace("init core lib %s", b->name);
-    if (b->open(l) != 1) {
-      logerr("failed to open library: %s", b->name);
-      return 0;
-    }
-    lua_setfield(l, -2, b->name);
-    trace("core set field: %s", b->name);
-  }
-  return 1;
 }
 
 int main(int argc, char *argv[]) {
